17/05/2021

+++ src/main/java/kasyan/springweb/service/RecoveryProductService.java
@@ -28,10 +28,13 @@ public class RecoveryProductService {
     //восстанавливаем удаленный ранее Product по его ID и отправка запроса в БД
     @Transactional
     public void recovered(int id) {
-        var productOfDelete = getProductService.findProductOfBascketByID(id);
-        saveProductService.saveProduct(productOfDelete.getCategory(), productOfDelete.getName(), productOfDelete.getPrice(),
-                productOfDelete.getDiscount(), productOfDelete.getTotalVolume());
-        deleteProductService.deleteOfBasket(id);
+        var productOfDeleted = getProductService.findProductOfBasketByID(id);
+        if (productOfDeleted.isPresent()) {
+            var productOfDelete = productOfDeleted.get();
+            saveProductService.saveProduct(productOfDelete.getCategory(), productOfDelete.getName(),
+                    productOfDelete.getPrice(), productOfDelete.getDiscount(), productOfDelete.getTotalVolume());
+            deleteProductService.deleteOfBasket(id);
+        }
     }
 
     @Autowired
